backend/requirements.txt
uvicorn
fastapi
sqlalchemy
psycopg2
pydantic
alembic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
python-multipart
